pyproject.toml
[project]
name = "mcp-server-creation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.1.0",
    "langchain-groq>=1.0.0",
    "langchain-mcp-adapters>=0.1.12",
    "langgraph>=1.0.2",
    "mcp>=1.20.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.28.0",
    "requests>=2.31.0",
]
